refactor: move local command playback scenario
