+++ tests/coding/test_native_tui_playback_runner.py
@@ -38,6 +38,7 @@ def test_native_tui_playback_fake_session_lists_command_sources() -> None:
 
 def test_native_tui_playback_command_scenarios_live_in_command_module() -> None:
     assert [scenario.name for scenario in COMMAND_ROUTING_SCENARIOS] == [
+        "local-command",
         "session-name-command",
         "session-command-error",
         "unknown-slash-prompt",
